feat: 컴포넌트 분리 및 반응형 레이아웃 추가 (#67)
